[FLINK-26995][tests] Migrate tests to ITCases
These classes were only renamed:
- FileUploadHandlerTest requires exclusive usage of DeleteOnExitHooks
- StreamTaskTimerTest requires exclusive access to to StreamTask#TRIGGER_THREAD_GROUP
- SinkTransformationTranslatorTest fails it
	a) relies on initially starting out with a clean state in the streaming graph generation
	b) relies on test cases permanently mutating the state in the streaming graph generation

Some tests of the following classes were moved into a separate ITCase:
- MemoryExecutionGraphInfoStoreTest relies on static SignallingBlockingNoOpInvokable
- StreamTaskTest requires exclusive usage of the OutputFlusher
