+++ flink-runtime/src/test/java/org/apache/flink/runtime/dispatcher/MemoryExecutionGraphInfoStoreTest.java
@@ -21,16 +21,9 @@ package org.apache.flink.runtime.dispatcher;
 import org.apache.flink.api.common.JobID;
 import org.apache.flink.api.common.JobStatus;
 import org.apache.flink.api.common.time.Time;
-import org.apache.flink.configuration.Configuration;
-import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.runtime.executiongraph.ArchivedExecutionGraph;
-import org.apache.flink.runtime.jobgraph.JobGraph;
-import org.apache.flink.runtime.jobgraph.JobGraphTestUtils;
-import org.apache.flink.runtime.jobgraph.JobVertex;
 import org.apache.flink.runtime.messages.webmonitor.JobDetails;
 import org.apache.flink.runtime.messages.webmonitor.JobsOverview;
-import org.apache.flink.runtime.minicluster.MiniCluster;
-import org.apache.flink.runtime.minicluster.MiniClusterConfiguration;
 import org.apache.flink.runtime.rest.handler.legacy.utils.ArchivedExecutionGraphBuilder;
 import org.apache.flink.runtime.scheduler.ExecutionGraphInfo;
 import org.apache.flink.runtime.util.ManualTicker;
@@ -237,29 +230,6 @@ public class MemoryExecutionGraphInfoStoreTest extends TestLogger {
             assertThat(
                     executionGraphInfoStore.getAvailableJobDetails(),
                     Matchers.containsInAnyOrder(jobDetails.toArray()));
-        }
-    }
-
-    /** Tests that a session cluster can terminate gracefully when jobs are still running. */
-    @Test
-    public void testPutSuspendedJobOnClusterShutdown() throws Exception {
-        Configuration configuration = new Configuration();
-        configuration.set(JobManagerOptions.JOB_STORE_TYPE, JobManagerOptions.JobStoreType.Memory);
-        try (final MiniCluster miniCluster =
-                new ExecutionGraphInfoStoreTestUtils.PersistingMiniCluster(
-                        new MiniClusterConfiguration.Builder()
-                                .setConfiguration(configuration)
-                                .build(),
-                        new ScheduledExecutorServiceAdapter(EXECUTOR_RESOURCE.getExecutor()))) {
-            miniCluster.start();
-            final JobVertex vertex = new JobVertex("blockingVertex");
-            // The adaptive scheduler expects that every vertex has a configured parallelism
-            vertex.setParallelism(1);
-            vertex.setInvokableClass(
-                    ExecutionGraphInfoStoreTestUtils.SignallingBlockingNoOpInvokable.class);
-            final JobGraph jobGraph = JobGraphTestUtils.streamingJobGraph(vertex);
-            miniCluster.submitJob(jobGraph);
-            ExecutionGraphInfoStoreTestUtils.SignallingBlockingNoOpInvokable.LATCH.await();
         }
     }
