+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/tasks/StreamTaskTest.java
@@ -25,7 +25,6 @@ import org.apache.flink.api.common.operators.ProcessingTimeService.ProcessingTim
 import org.apache.flink.api.common.state.CheckpointListener;
 import org.apache.flink.api.common.typeinfo.BasicTypeInfo;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
-import org.apache.flink.api.common.typeutils.base.StringSerializer;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.ReadableConfig;
 import org.apache.flink.core.execution.SavepointFormatType;
@@ -45,7 +44,6 @@ import org.apache.flink.runtime.checkpoint.SnapshotType;
 import org.apache.flink.runtime.checkpoint.SubtaskState;
 import org.apache.flink.runtime.checkpoint.TaskStateSnapshot;
 import org.apache.flink.runtime.checkpoint.channel.ChannelStateWriter;
-import org.apache.flink.runtime.deployment.ResultPartitionDeploymentDescriptor;
 import org.apache.flink.runtime.execution.CancelTaskException;
 import org.apache.flink.runtime.execution.Environment;
 import org.apache.flink.runtime.execution.ExecutionState;
@@ -66,7 +64,6 @@ import org.apache.flink.runtime.operators.testutils.DummyEnvironment;
 import org.apache.flink.runtime.operators.testutils.ExpectedTestException;
 import org.apache.flink.runtime.operators.testutils.MockEnvironment;
 import org.apache.flink.runtime.operators.testutils.MockEnvironmentBuilder;
-import org.apache.flink.runtime.shuffle.PartitionDescriptorBuilder;
 import org.apache.flink.runtime.shuffle.ShuffleEnvironment;
 import org.apache.flink.runtime.state.AbstractKeyedStateBackend;
 import org.apache.flink.runtime.state.AbstractStateBackend;
@@ -102,7 +99,6 @@ import org.apache.flink.runtime.taskmanager.TaskExecutionState;
 import org.apache.flink.runtime.taskmanager.TaskManagerActions;
 import org.apache.flink.runtime.taskmanager.TestTaskBuilder;
 import org.apache.flink.runtime.throughput.ThroughputCalculator;
-import org.apache.flink.runtime.util.NettyShuffleDescriptorBuilder;
 import org.apache.flink.runtime.util.TestingTaskManagerRuntimeInfo;
 import org.apache.flink.streaming.api.TimeCharacteristic;
 import org.apache.flink.streaming.api.checkpoint.CheckpointedFunction;
@@ -110,15 +106,12 @@ import org.apache.flink.streaming.api.functions.source.RichParallelSourceFunctio
 import org.apache.flink.streaming.api.functions.source.SourceFunction;
 import org.apache.flink.streaming.api.graph.StreamConfig;
 import org.apache.flink.streaming.api.operators.AbstractStreamOperator;
-import org.apache.flink.streaming.api.operators.AbstractStreamOperatorFactory;
-import org.apache.flink.streaming.api.operators.ChainingStrategy;
 import org.apache.flink.streaming.api.operators.InternalTimeServiceManager;
 import org.apache.flink.streaming.api.operators.OneInputStreamOperator;
 import org.apache.flink.streaming.api.operators.OperatorSnapshotFutures;
 import org.apache.flink.streaming.api.operators.Output;
 import org.apache.flink.streaming.api.operators.StreamMap;
 import org.apache.flink.streaming.api.operators.StreamOperator;
-import org.apache.flink.streaming.api.operators.StreamOperatorParameters;
 import org.apache.flink.streaming.api.operators.StreamOperatorStateContext;
 import org.apache.flink.streaming.api.operators.StreamSource;
 import org.apache.flink.streaming.api.operators.StreamTaskStateInitializer;
@@ -134,17 +127,14 @@ import org.apache.flink.util.CloseableIterable;
 import org.apache.flink.util.ExceptionUtils;
 import org.apache.flink.util.FatalExitExceptionHandler;
 import org.apache.flink.util.FlinkRuntimeException;
-import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.TestLogger;
 import org.apache.flink.util.clock.SystemClock;
 import org.apache.flink.util.concurrent.FutureUtils;
 import org.apache.flink.util.concurrent.TestingUncaughtExceptionHandler;
 import org.apache.flink.util.function.BiConsumerWithException;
-import org.apache.flink.util.function.FunctionWithException;
 import org.apache.flink.util.function.RunnableWithException;
 import org.apache.flink.util.function.SupplierWithException;
 
-import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matchers;
 import org.junit.Assert;
 import org.junit.ClassRule;
@@ -183,7 +173,6 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Consumer;
 
 import static java.util.Arrays.asList;
-import static java.util.Collections.singletonList;
 import static org.apache.flink.api.common.typeinfo.BasicTypeInfo.STRING_TYPE_INFO;
 import static org.apache.flink.configuration.StateBackendOptions.STATE_BACKEND;
 import static org.apache.flink.configuration.TaskManagerOptions.BUFFER_DEBLOAT_ENABLED;
@@ -193,15 +182,12 @@ import static org.apache.flink.configuration.TaskManagerOptions.BUFFER_DEBLOAT_T
 import static org.apache.flink.configuration.TaskManagerOptions.MEMORY_SEGMENT_SIZE;
 import static org.apache.flink.runtime.checkpoint.CheckpointFailureReason.UNKNOWN_TASK_CHECKPOINT_NOTIFICATION_FAILURE;
 import static org.apache.flink.runtime.checkpoint.StateObjectCollection.singleton;
-import static org.apache.flink.runtime.io.network.api.writer.RecordWriter.DEFAULT_OUTPUT_FLUSH_THREAD_NAME;
 import static org.apache.flink.runtime.state.CheckpointStorageLocationReference.getDefault;
 import static org.apache.flink.streaming.runtime.tasks.mailbox.TaskMailbox.MAX_PRIORITY;
 import static org.apache.flink.streaming.util.StreamTaskUtil.waitTaskIsRunning;
 import static org.apache.flink.util.Preconditions.checkState;
 import static org.hamcrest.CoreMatchers.both;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.not;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
@@ -1184,110 +1170,6 @@ public class StreamTaskTest extends TestLogger {
     }
 
     @Test
-    public void testRecordWriterClosedOnTransitDeployingStateError() throws Exception {
-        testRecordWriterClosedOnTransitStateError(ExecutionState.DEPLOYING);
-    }
-
-    @Test
-    public void testRecordWriterClosedOnTransitInitializingStateError() throws Exception {
-        testRecordWriterClosedOnTransitStateError(ExecutionState.INITIALIZING);
-    }
-
-    @Test
-    public void testRecordWriterClosedOnTransitRunningStateError() throws Exception {
-        testRecordWriterClosedOnTransitStateError(ExecutionState.RUNNING);
-    }
-
-    private void testRecordWriterClosedOnTransitStateError(ExecutionState executionState)
-            throws Exception {
-        // Throw the exception when the state updating to the expected one.
-        NoOpTaskManagerActions taskManagerActions =
-                new NoOpTaskManagerActions() {
-                    @Override
-                    public void updateTaskExecutionState(TaskExecutionState taskExecutionState) {
-                        if (taskExecutionState.getExecutionState() == executionState) {
-                            throw new ExpectedTestException();
-                        }
-                    }
-                };
-
-        testRecordWriterClosedOnError(
-                env ->
-                        taskBuilderWithConfiguredRecordWriter(env)
-                                .setTaskManagerActions(taskManagerActions)
-                                .build(EXECUTOR_RESOURCE.getExecutor()));
-    }
-
-    @Test
-    public void testFailInEndOfConstructor() throws Exception {
-        Configuration conf = new Configuration();
-        // Set the wrong setting type for forcing the fail during read.
-        conf.setString(BUFFER_DEBLOAT_PERIOD.key(), "a");
-        testRecordWriterClosedOnError(
-                env ->
-                        taskBuilderWithConfiguredRecordWriter(env)
-                                .setTaskManagerConfig(conf)
-                                .build(EXECUTOR_RESOURCE.getExecutor()));
-    }
-
-    private void testRecordWriterClosedOnError(
-            FunctionWithException<NettyShuffleEnvironment, Task, Exception> taskProvider)
-            throws Exception {
-        try (NettyShuffleEnvironment shuffleEnvironment =
-                new NettyShuffleEnvironmentBuilder().build()) {
-            Task task = taskProvider.apply(shuffleEnvironment);
-
-            task.startTaskThread();
-            task.getExecutingThread().join();
-
-            assertEquals(ExecutionState.FAILED, task.getExecutionState());
-            for (Thread thread : Thread.getAllStackTraces().keySet()) {
-                assertThat(
-                        thread.getName(),
-                        CoreMatchers.is(not(containsString(DEFAULT_OUTPUT_FLUSH_THREAD_NAME))));
-            }
-        }
-    }
-
-    private TestTaskBuilder taskBuilderWithConfiguredRecordWriter(
-            NettyShuffleEnvironment shuffleEnvironment) {
-        Configuration taskConfiguration = new Configuration();
-        outputEdgeConfiguration(taskConfiguration);
-
-        ResultPartitionDeploymentDescriptor descriptor =
-                new ResultPartitionDeploymentDescriptor(
-                        PartitionDescriptorBuilder.newBuilder().build(),
-                        NettyShuffleDescriptorBuilder.newBuilder().buildLocal(),
-                        1,
-                        false);
-        return new TestTaskBuilder(shuffleEnvironment)
-                .setInvokable(NoOpStreamTask.class)
-                .setTaskConfig(taskConfiguration)
-                .setResultPartitions(singletonList(descriptor));
-    }
-
-    /**
-     * Make sure that there is some output edge in the config so that some RecordWriter is created.
-     */
-    private void outputEdgeConfiguration(Configuration taskConfiguration) {
-        StreamConfig streamConfig = new StreamConfig(taskConfiguration);
-        streamConfig.setStreamOperatorFactory(new UnusedOperatorFactory());
-
-        StreamConfigChainer cfg =
-                new StreamConfigChainer(new OperatorID(42, 42), streamConfig, this, 1);
-        // The OutputFlusher thread is started only if the buffer timeout more than 0(default value
-        // is 0).
-        cfg.setBufferTimeout(1);
-        cfg.chain(
-                new OperatorID(44, 44),
-                new UnusedOperatorFactory(),
-                StringSerializer.INSTANCE,
-                StringSerializer.INSTANCE,
-                false);
-        cfg.finish();
-    }
-
-    @Test
     public void testProcessWithAvailableOutput() throws Exception {
         try (final MockEnvironment environment = setupEnvironment(true, true)) {
             final int numberOfProcessCalls = 10;
@@ -1998,27 +1880,6 @@ public class StreamTaskTest extends TestLogger {
     }
 
     /**
-     * Operator that does nothing.
-     *
-     * @param <T>
-     * @param <OP>
-     */
-    public static class NoOpStreamTask<T, OP extends StreamOperator<T>> extends StreamTask<T, OP> {
-
-        public NoOpStreamTask(Environment environment) throws Exception {
-            super(environment);
-        }
-
-        @Override
-        protected void init() throws Exception {
-            inputProcessor = new EmptyInputProcessor();
-        }
-
-        @Override
-        protected void cleanUpInternal() throws Exception {}
-    }
-
-    /**
      * A stream input processor implementation used to control the returned input status based on
      * the total number of processing calls.
      */
@@ -2080,31 +1941,6 @@ public class StreamTaskTest extends TestLogger {
         @Override
         public CompletableFuture<?> getAvailableFuture() {
             return availabilityProvider.getAvailableFuture();
-        }
-    }
-
-    private static class BlockingFinishStreamOperator extends AbstractStreamOperator<Void> {
-        private static final long serialVersionUID = -9042150529568008847L;
-
-        private static volatile OneShotLatch inFinish;
-        private static volatile OneShotLatch finishClose;
-
-        @Override
-        public void finish() throws Exception {
-            checkLatches();
-            inFinish.trigger();
-            finishClose.await();
-            super.close();
-        }
-
-        private void checkLatches() {
-            Preconditions.checkNotNull(inFinish);
-            Preconditions.checkNotNull(finishClose);
-        }
-
-        private static void resetLatches() {
-            inFinish = new OneShotLatch();
-            finishClose = new OneShotLatch();
         }
     }
 
@@ -2241,7 +2077,7 @@ public class StreamTaskTest extends TestLogger {
         }
     }
 
-    private static class EmptyInputProcessor implements StreamInputProcessor {
+    static class EmptyInputProcessor implements StreamInputProcessor {
         private volatile boolean isFinished;
 
         public EmptyInputProcessor() {
@@ -2685,22 +2521,6 @@ public class StreamTaskTest extends TestLogger {
         @Override
         public void failExternally(Throwable cause) {
             throw failingCause;
-        }
-    }
-
-    private static class UnusedOperatorFactory extends AbstractStreamOperatorFactory<String> {
-        @Override
-        public <T extends StreamOperator<String>> T createStreamOperator(
-                StreamOperatorParameters<String> parameters) {
-            throw new UnsupportedOperationException("This shouldn't be called");
-        }
-
-        @Override
-        public void setChainingStrategy(ChainingStrategy strategy) {}
-
-        @Override
-        public Class<? extends StreamOperator> getStreamOperatorClass(ClassLoader classLoader) {
-            throw new UnsupportedOperationException();
         }
     }
